requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
firebase-admin==6.2.0
motor==3.3.2
pymongo==4.6.0
aiofiles==23.2.1
python-multipart==0.0.6
